Fixed icons in custom dropdowns (#5041)
* Fixed vseparation references to v_separation

* Reduced temporary allocations in CustomDropDown.cs

* Fixed auto-evo exploring tool report not displaying

* Fixed custom drop down icon texture drawing

by swapping to the Godot 4 inbuilt support for it

* Fixed mucilage emission not working for the player

* Thriveopedia organelle link button text is now center aligned

+++ src/microbe_stage/PlayerMicrobeInput.cs
@@ -129,14 +129,14 @@ public partial class PlayerMicrobeInput : NodeWithInput
     }
 
     [RunOnKey("g_secrete_slime")]
-    public void SecreteSlime(float delta)
+    public void SecreteSlime(double delta)
     {
         if (!stage.HasPlayer)
             return;
 
         ref var control = ref stage.Player.Get<MicrobeControl>();
 
-        control.QueueSecreteSlime(ref stage.Player.Get<OrganelleContainer>(), stage.Player, delta);
+        control.QueueSecreteSlime(ref stage.Player.Get<OrganelleContainer>(), stage.Player, (float)delta);
     }
 
     [RunOnKeyDown("g_toggle_engulf")]
